Changed Locations, Packaged, Debugged
I changed the locations of the files to be consistent of what was going on in the project. I also packaged the MAUI framework, but have only included the most recent one. Futhermore, I have spent much time debugging this since I needed a child SMO (MAUI Template) to really work out the bugs.